+++ messages/definitions/condition-messages.ts
@@ -49,7 +49,7 @@ export interface IMaxRepetitionsCondition {
 
 export const Repetitions = [
     { name: "enabled", type: "bool" },                    // indicates whether the condition should be checked
-    { name: "amount", type: "uint16" },                 // the maximum number of times the script can be ran
+    { name: "amount", type: "uint32" },                 // the maximum number of times the script can be ran
 ];
 
 export interface IFollowCondition {
